+++ ShopApp/Controllers/OrderController.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using ShopApp.Application.Interfaces.Order;
+using ShopApp.Domain.Common;
 using ShopApp.Domain.DTOs.Order;
 
 namespace ShopApp.Controllers
@@ -9,10 +11,12 @@ namespace ShopApp.Controllers
     public class OrderController : ControllerBase
     {
         private readonly IOrderService _orderService;
+        private IAuthorizationService _authorizationService;
 
-        public OrderController(IOrderService orderService)
+        public OrderController(IOrderService orderService, IAuthorizationService authorizationService)
         {
             _orderService = orderService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet("{id}")]
@@ -40,6 +44,9 @@ namespace ShopApp.Controllers
         [HttpPost]
         public async Task<ActionResult> Add([FromBody] AddOrderRequestDto newOrder)
         {
+            if (await Authorize(newOrder) is false)
+                return Forbid();
+
             var orderResponse = await _orderService.Add(newOrder);
             return Ok(orderResponse);
         }
@@ -47,6 +54,11 @@ namespace ShopApp.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete([FromRoute] int id)
         {
+            var getOrder = await _orderService.GetById(id);
+
+            if (await Authorize(getOrder) is false)
+                return Forbid();
+
             var order = await _orderService.Delete(id);
 
             if (order is null)
@@ -58,12 +70,30 @@ namespace ShopApp.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult> Update([FromRoute] int id, AddOrderRequestDto newOrder)
         {
+            var getOrder = await _orderService.GetById(id);
+
+            if (await Authorize(getOrder) is false)
+                return Forbid();
+
             var order = await _orderService.Update(id, newOrder);
 
             if (order is null)
                 return NotFound(order);
 
             return Ok(order);
+        }
+
+        [NonAction]
+        public async Task<bool> Authorize(BaseCustomer baseCustomer)
+        {
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, baseCustomer, "CustomerPolicy");
+
+            if (!authorizationResult.Succeeded)
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
